remove rubocop eroors
